coefTable {miscTools}	R Documentation

Coefficient Table

Description

Generate Table for Coefficients, Std. Errors, t-values and P-values.

Usage

   coefTable( coef, stdErr, df = NULL )

Arguments

coef
vector that contains the coefficients.

stdErr
vector that contains the standard errors of the coefficients.

df
degrees of freedom of the t-test used to calculate P-values.

Value

a matrix with 4 columns: coefficients, standard errors, t-values and P-values. If argument df is not provided, the last column (P-values) is filled with NAs.

Author(s)

Arne Henningsen

Examples

   coefTable( rnorm( 10 ), 0.5 * abs( rnorm( 10 ) ), 20 )

[Package miscTools version 0.6-30 Index]