ddnorm {miscTools}	R Documentation

Derivative of the Normal Distribution's Density Function

Description

This function returns the derivative(s) of the density function of the normal (Gaussian) distribution with respect to the quantile, evaluated at the quantile(s), mean(s), and standard deviation(s) specified by arguments x, mean, and sd, respectively.

Usage

ddnorm( x, mean = 0, sd = 1 )

Arguments

x
quantile or vector of quantiles.

mean
mean or vector of means.

sd
standard deviation or vector of standard deviations.

Value

numeric value(s): derivative(s) of the density function of the normal distribution with respect to the quantile

Author(s)

Arne Henningsen

See Also

dnorm

Examples

ddnorm( c( -1, 0, 1 ) )

[Package miscTools version 0.6-30 Index]