insertCol {miscTools}	R Documentation

Insert Column into a Matrix

Description

Insert a new column into a matrix.

Usage

insertCol( m, c, v = NA, cName = "" )

Arguments

m
matrix.

c
column number where the new column should be inserted.

v
optional values of the new column.

cName
optional character string: the name of the new column.

Value

a matrix with one more column than the provided matrix m.

Author(s)

Arne Henningsen

See Also

insertRow.

Examples

   m <- matrix( 1:4, 2 )
   insertCol( m, 2, 5:6 )

[Package miscTools version 0.6-30 Index]