insertRow {miscTools}	R Documentation

Insert Row into a Matrix

Description

Insert a new row into a matrix.

Usage

insertRow( m, r, v = NA, rName = "" )

Arguments

m
matrix.

r
row number where the new row should be inserted.

v
optional values for the new row.

rName
optional character string: the name of the new row.

Value

a matrix with one more row than the provided matrix m.

Author(s)

Arne Henningsen

See Also

insertCol.

Examples

   m <- matrix( 1:4, 2 )
   insertRow( m, 2, 5:6 )

[Package miscTools version 0.6-30 Index]