margEff {miscTools}	R Documentation

Method for Returning Marginal Effects

Description

Currently, this package just defines the generic function margEff so that it can be used to define margEff methods for objects of specific classes in other packages.

Usage

margEff( object, ... )

Arguments

object
an object of which marginal effects should be calculated.

...
further arguments for methods

Author(s)

Arne Henningsen

[Package miscTools version 0.6-30 Index]