quasiconcavity {miscTools}	R Documentation

Test for quasiconcavity / quasiconvexity

Description

Test wether a function is quasiconcave or quasiconvex. The bordered Hessian of this function is checked by quasiconcavity() or quasiconvexity().

Usage

   quasiconcavity( m, tol = .Machine$double.eps )
   quasiconvexity( m, tol = .Machine$double.eps )

Arguments

m
a bordered Hessian matrix or a list containing bordered Hessian matrices

tol
tolerance level (values between -tol and tol are considered to be zero).

Value

locigal or a logical vector (if m is a list).

Author(s)

Arne Henningsen

References

Chiang, A.C. (1984) Fundamental Methods of Mathematical Economics, 3rd ed., McGraw-Hill.

Examples

   quasiconcavity( matrix( 0, 3, 3 ) )

   quasiconvexity( matrix( 0, 3, 3 ) )

   m <- list()
   m[[1]] <- matrix( c( 0,-1,-1, -1,-2,3, -1,3,5 ), 3, 3 )
   m[[2]] <- matrix( c( 0,1,-1, 1,-2,3, -1,3,5 ), 3, 3 )

   quasiconcavity( m )

   quasiconvexity( m )

[Package miscTools version 0.6-30 Index]