rowMedians {miscTools}	R Documentation

Medians of Rows

Description

Compute the sample medians of the rows of a data.frame or matrix.

Usage

rowMedians( x, na.rm = FALSE )

Arguments

x
a data.frame or matrix.

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

Value

A vector of the medians of each row of x.

Author(s)

Arne Henningsen

See Also

colMedians,median,colMeans.

Examples

   m <- matrix( 1:12, nrow = 4 )
   rowMedians( m )

[Package miscTools version 0.6-30 Index]