stdEr {miscTools}	R Documentation

Standard deviations

Description

Extract standard deviations from estimated models.

Usage

  stdEr(x, ...)
## Default S3 method:
stdEr(x, ...)
## S3 method for class 'lm'
stdEr(x, ...)

Arguments

x
a statistical model, such as created by lm

...
further arguments for methods

Details

stdEr is a generic function with methods for objects of "lm" class. The default method returns the square root of the diagonal of the variance-covariance matrix.

Value

numeric, the estimated standard errors of the coefficients.

Author(s)

Ott Toomet otoomet@ut.ee

See Also

vcov, summary.

Examples

data(cars)
lmRes <- lm(dist ~ speed, data=cars)
stdEr( lmRes )

[Package miscTools version 0.6-30 Index]