summarizeDF {miscTools}	R Documentation

Summarize a data.rrame

Description

This function summarizes each variable that is in a data.frame. It can be used, e.g., in an R script to write summary information about a data.frame into a text file that is in a version control system so that one can see in the version control system whether one or more variables in the data frame have changed.

Usage

summarizeDF( dat, printValues = TRUE, maxLevel = 20, file = NULL, ... )

Arguments

dat
a data.frame.

printValues
logical. If FALSE only MD5 checksums are returned, which could be desirable if the data frame contains confidential data that should not be included in the output.

maxLevel
integer. If the number of unique values in a variable is less than or equal to the number specified in this argument (and argument printValues is TRUE), a frequency table is included in the output.

file
a character string or a writable connection naming the file to write to.

...
further arguments forwarded to sink() if argument file is not NULL.

Author(s)

Arne Henningsen

[Package miscTools version 0.6-30 Index]