vecli2m {miscTools}	R Documentation

Convert vector of linear independent values into a Matrix

Description

Converts a vector into a symmetric matrix that the original vector contains the linear independent values of the returned symmetric matrix.

Usage

vecli2m( v )

Arguments

v
a vector.

Author(s)

Arne Henningsen

See Also

vecli, veclipos.

Examples

   v <- c( 11, 12, 13, 22, 23, 33 )
   vecli2m( v )

[Package miscTools version 0.6-30 Index]