veclipos {miscTools}	R Documentation

Position in a vector of linear independent values

Description

Returns the position of the [i,j]th element of a symmetric n x n matrix that this element has in a vector of the linear independent values of the matrix.

Usage

veclipos( i, j, n )

Arguments

i
row of the element in the matrix.

j
column of the element in the matrix.

n
dimension of the matrix.

Note

A symmetric n x n matrix has n*(n+1)/2 independent values.
The function is: n*(n-1)/2-((n-min(i,j))*(n-min(i,j)+1)/2)+max(i,j)

Author(s)

Arne Henningsen

See Also

vecli, vecli2m.

Examples

   veclipos( 1, 2, 3 )  # returns: 2

[Package miscTools version 0.6-30 Index]